Removed variants from icon, because it cause problems in the button. Removed align on label in button after refactoring. Fixed loading and disabled bugs on label in the button.
