+++ src/components/Button/Button.css
@@ -4,9 +4,15 @@
 .button:disabled {
   @apply opacity-50 cursor-not-allowed;
 }
+.button:disabled > * {
+  @apply cursor-not-allowed;
+}
 
 .button--loading {
   @apply animate-pulse cursor-wait;
+}
+.button--loading > * {
+  @apply cursor-wait;
 }
 
 .button--xl {
@@ -37,7 +43,6 @@
 .button--variant-secondary {
   @apply bg-white border border-neutral-100 hover:border-neutral-700 text-medium-900 hover:text-dark-500 focus:outline-none focus:ring-4 focus:ring-accent-200;
 }
-
 .button--variant-text {
   @apply hover:bg-white border border-transparent hover:border-neutral-200 text-medium-900 focus:outline-none focus:ring-4 focus:ring-accent-200;
 }
