+++ src/components/Icon/Icon.stories.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Icon } from "../Icon";
+import { Icon } from "./Icon";
 import { HiBadgeCheck } from "react-icons/hi";
 
 export default {
